# Optimize data properties