--- a/src/lib/byte-length-queuing-strategy.ts
+++ b/src/lib/byte-length-queuing-strategy.ts
@@ -1,11 +1,10 @@
-import { createDataProperty } from './helpers';
 import { QueuingStrategy } from './queuing-strategy';
 
 export default class ByteLengthQueuingStrategy implements QueuingStrategy<ArrayBufferView> {
   readonly highWaterMark!: number;
 
   constructor({ highWaterMark }: { highWaterMark: number }) {
-    createDataProperty(this, 'highWaterMark', highWaterMark);
+    this.highWaterMark = highWaterMark;
   }
 
   size(chunk: ArrayBufferView): number {