--- a/src/lib/count-queuing-strategy.ts
+++ b/src/lib/count-queuing-strategy.ts
@@ -1,11 +1,10 @@
-import { createDataProperty } from './helpers';
 import { QueuingStrategy } from './queuing-strategy';
 
 export default class CountQueuingStrategy implements QueuingStrategy<any> {
   readonly highWaterMark!: number;
 
   constructor({ highWaterMark }: { highWaterMark: number }) {
-    createDataProperty(this, 'highWaterMark', highWaterMark);
+    this.highWaterMark = highWaterMark;
   }
 
   size(): 1 {